requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
djangorestframework==3.9.0
Flask==1.0.2
future==0.17.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyee==5.0.0
python-twitter==3.5
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
slackclient==1.3.0
slackeventsapi==2.0.0
urllib3==1.24
websocket-client==0.53.0
Werkzeug==0.14.1
whitenoise==4.1
